+++ src/async/Services/index.js
@@ -1,3 +1,4 @@
 module.exports.http = require("./Http");
+module.exports.graphQL = require("./GraphQL");
 module.exports.serializer = require("./Serializer");
 module.exports.Trait = require("./Trait");