+++ CHANGELOG.md
@@ -15,6 +15,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 
 ### Changed
 
+- Changed scheduling requests that pass now through Alfred (GraphQL API).
+
 ### Deprecated
 
 ### Removed